Moved HOSTTYPE defs to environment.h since they also need to be used in xlogin.c. (These also contained an ifdef sparc which I changed to SOLARIS for consistency and truth.)
Added /srvd/patch to defaultpath, as it really ought to be there.
There may be other problems with defaultpath, I don't know.

Changed LD_LIBRARY_PATH to just pass out of xlogin the value that
is passed in by dm. Added OPENWINHOME to do the same thing. Added
a macro to make this convenient to do, and changed TZ to use it
as well.

Added comments explaining xlogin's general environment setup 
scheme.