Eliminate space before #else in 1.71.
The consequences of this extra space were sort of interesting.  When
this file is copied to athena-verify.c and built for /etc/athena/xdm,
it's built with the ANSI C preprocessor (no -traditional flag to gcc),
and spaces before pound signs aren't a problem.  When this file is
built for xlogin, it's built with the K&R C preprocessor (gcc
-traditional), and spaces before pound signs are a problem--but XDM
isn't defined, so the effect is to do nothing at all with OPENWINHOME.
So OPENWINHOME never gets set and programs can't find XKeysymDB.